pyproject.toml
[project]
name = "essential-model-documentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandoc>=2.4",
    "python-docx>=1.1.0",
]
